requirements.txt
torch
torchvision
h5py
flask
groq
python-dotenv
matplotlib
ipykernel
tensorflow